pyproject.toml
[project]
name = "mnist-pure-python"
version = "0.1.0"
description = "MNIST without any ML frameworks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.2",
    "pillow>=11.3.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pre-commit>=4.3.0",
]
